requirements.txt
flask
pytesseract
pdf2image
pillow
requests
langchain_openai
langchain_core
gunicorn
